refactor(audit): split drift renderers after rebase
Merged cold-cache replay growth pushed install/drift.py past the 1000-line architecture budget. Move output-only rendering to drift_render.py while re-exporting the existing API, preserving both #2329 replay ownership and #2380 membership semantics.
